+++ src/index.ts
@@ -9,6 +9,7 @@ export { isArray } from './is_array';
 export { identity } from './identity';
 export { tryCatch } from './try_catch';
 export { sequence } from './sequence';
+export { arrayFrom } from './array_from';
 export { promiseOf } from './promise_of';
 export { isNothing } from './is_nothing';
 export { isPromise } from './is_promise';