+++ test/array_from.test.ts
@@ -1,0 +1,17 @@
+import { arrayFrom } from '../src';
+
+describe('arrayFrom', () => {
+  test('arrayFrom creates an array from iterables', () => {
+    expect(typeof arrayFrom([9], new Set([6]), { 0: 6, length: 1 })).toBe(
+      'object'
+    );
+    expect(
+      Array.isArray(arrayFrom([9], new Set([6]), { 0: 6, length: 1 }))
+    ).toBe(true);
+    expect(arrayFrom([9], new Set([6]), { 0: 6, length: 1 })).toEqual([
+      9,
+      6,
+      6,
+    ]);
+  });
+});