+++ src/declaration/index.d.ts
@@ -285,6 +285,28 @@ declare module '@fluss/core' {
     ...iterables: ReadonlyArray<ArrayLike<T> | Iterable<T>>
   ): ReadonlyArray<T>;
 
+  /** Creates readonly type from set of elements. */
+  export function tupleOf<V>(v: V): readonly [V];
+  export function tupleOf<V, V1>(v: V, v1: V1): readonly [V, V1];
+  export function tupleOf<V, V1, V2>(
+    v: V,
+    v1: V1,
+    v2: V2
+  ): readonly [V, V1, V2];
+  export function tupleOf<V, V1, V2, V3>(
+    v: V,
+    v1: V1,
+    v2: V2,
+    v3: V3
+  ): readonly [V, V1, V2, V3];
+  export function tupleOf<V, V1, V2, V3, V4>(
+    v: V,
+    v1: V1,
+    v2: V2,
+    v3: V3,
+    v4: V4
+  ): readonly [V, V1, V2, V3, V4];
+
   /**
    * Wraps code into `try/catch` and returns `Either` monad with result.
    * If `catchFn` is not `undefined`, then `Either` with result will